listapi/helpdesk/ - replace remove.Image.php to remove.File.php

+++ include/listapi/helpdesk/remove.File.php
@@ -11,7 +11,7 @@ try{
 	
 	/* useExam
 	
-		dataReqNext({file:'listapi/helpdesk/remove.Image.php',type:'json',args:'dataRecord={"nimage":53}'},
+		dataReqNext({file:'listapi/helpdesk/remove.File.php',type:'json',args:'dataRecord={"nimage":53}'},
 			console.log
 		);
 	*/
@@ -28,12 +28,12 @@ try{
 	/* availabla for use $dataRecord */
 	
 	if(empty($dataRecord['nimage'])){
-		throw new Exception("arg nimage is empty (for remove.Image)");
+		throw new Exception("arg nimage is empty (for remove.File)");
 	}
 	
 	$nimage = (int)$dataRecord['nimage'];
 	if($nimage <= 0){
-		throw new Exception("arg nimage is invalid (for remove.Image)");
+		throw new Exception("arg nimage is invalid (for remove.File)");
 	}
 	
 	// $result = $db->update("users",array(),array('avatar'=>""),"id=".$user['id']);
@@ -42,19 +42,18 @@ try{
 	$query = "SELECT i.image,i.path,ic.ncontext FROM images as i LEFT JOIN images_context as ic ON ic.nimage = i.id WHERE type=1 AND ic.nimage=".$nimage;
 	$result = $db->fetchFirst($query, $uid);
 	if( is_string($result) && (strpos($result,"error") !== false)){
-		throw new Exception("SQL Error (for remove.Image)");
+		throw new Exception("SQL Error (for remove.File)");
 	}
 	
 	if((!is_array($result)) || (count($result) == 0)){
-		throw new Exception("Data image is not found (for remove.Image)");
+		throw new Exception("Data image is not found (for remove.File)");
 	}
 	
 	$image = $result['image'];
 	$pathImage = $result['path'];
 	$nrequest = $result['ncontext'];
 	
-	
-	if(!$admin){
+	/* if(!$admin){ */
 		/* теперь рассмотреть права пользователя отновительно этой заявки */
 		/* выбираем данные по заявке $nrequest */
 		require_once $path."lib/query/get.query.ListTickets.php";
@@ -66,30 +65,45 @@ try{
 		/* проверяем права авторизованного */
 		
 		/* являемся ли мы инициатором. а Ответственный - он имеет права админа, а когда он админ - ему можно удалять любые изображения заявок */
-		if($dataTickets['applicant'] != $user['uid']){
+		if((!$admin) && ($dataTickets['applicant'] != $user['uid'])){
 			header("HTTP/1.1 403 Forbidden");
 			exit;
 		}
-	}
+		
+		if($dataTickets['hidden'] == 1){
+			header("HTTP/1.1 403 Forbidden");
+			exit;
+		}
+		
+		
+	/* } */
 	
 	// throw new ErrorException($_SERVER['PHP_SELF']);
 	//throw new ErrorException($_SERVER['DOCUMENT_ROOT']);
 	
 	
+	/* теперь узнать, не встречается ли это имя файла в других записях images, и если нет - то удаляем физически это изображение  */
+	$query = "SELECT i.image,i.path,ic.ncontext FROM images as i LEFT JOIN images_context as ic ON ic.nimage = i.id WHERE i.image='$image' AND ic.nimage<>".$nimage;
+	$result = $db->fetchFirst($query, $uid);
+	if( is_string($result) && (strpos($result,"error") !== false)){
+		throw new Exception("SQL Error (for remove.File)");
+	}
 	
-	/* это имя файла, который нужно удалить */
-	if(!unlink($_SERVER['DOCUMENT_ROOT'].$pathImage.$image)){
-		throw new Exception("RemoveFile Error (for remove.Image)");
+	if(count($result) == 0){
+		/* это имя файла, который нужно удалить */
+		if(!unlink($_SERVER['DOCUMENT_ROOT'].$pathImage.$image)){
+			throw new Exception("RemoveFile Error (for remove.File)");
+		}
 	}
 	
 	/* type=1 это изображения к заявкам */
 	$query = "DELETE FROM images_context WHERE type=1 AND nimage=".$nimage;
 	$result = $db->query($query, true, true);
 	if( is_string($result) && (strpos($result,"error") !== false)){
-		throw new Exception("SQL Error (for remove.Image)");
+		throw new Exception("SQL Error (for remove.File)");
 	}
 	
-	echo '{"success":1,"description":"изображение удалено","data":"'.$image.'"}';
+	echo '{"success":1,"description":"файл удален","data":"'.$image.'"}';
 	
 	
 }catch(ErrorException $ex){
